requirements.txt
numpy
pandas
matplotlib
Pillow
tensorflow-gpu==2.5.0
opencv-python==4.5.5
scikit-learn
patchify